requirements.txt
frida
capstone
pefile
pyelftools
anthropic
textual>=0.52.0
rich>=13.0.0
